loading items for spesific user completed

+++ src/components/ManageItems/ManageItems.js
@@ -3,31 +3,56 @@ import { Button, Table } from 'react-bootstrap';
 import { Link } from 'react-router-dom';
 import useBikes from './../../hooks/useBikes'
 import './ManageItems.css';
+import { PlusIcon, TrashIcon } from '@heroicons/react/solid'
+import { useAuthState } from 'react-firebase-hooks/auth';
+import auth from '../../firebase.init';
 
 const ManageItems = () => {
     //accessing all bikes from hook
     const [bikes, setBikes] = useBikes();
+    const [user] = useAuthState(auth);
 
     //handling delete
     const removeItem = (id) => {
-        console.log(id);
+        const confirm = window.confirm('Are you sure you want to delete?');
         //removing item from database
-        fetch(`http://localhost:5000/product/${id}`, {
-            method: 'DELETE'
+        if (confirm) {
+            fetch(`http://localhost:5000/product/${id}`, {
+                method: 'DELETE'
+            })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount) {
+                        // console.log('deleted from ui too');
+                        //filter other bikes and setting up a new state
+                        const remaining = bikes.filter(user => user._id !== id);
+                        setBikes(remaining);
+                    }
+                })
+        }
+    }
+    const addToOrder = (bike) => {
+        const email = user.email;
+        const img = bike.img;
+        const name = bike.name;
+        const supplierName = bike.supplierName;
+        const newBike = { img, name, supplierName, email }
+        console.log(newBike);
+        fetch(`http://localhost:5000/add-order`, {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify(newBike)
         })
             .then(res => res.json())
             .then(data => {
-                if (data.deletedCount) {
-                    console.log('deleted');
-                    //filter other bikes and setting up a new state
-                    const remaining = bikes.filter(user => user._id !== id);
-                    setBikes(remaining);
-                }
+                console.log(data);
             })
     }
     return (
         <div>
-            {/* {bikes.length} */}
+
             <div className='container'>
 
                 <Button as={Link} to='/add-item' className='mt-5 d-block w-25' style={{ marginLeft: '70%' }}> Add item</Button>
@@ -38,14 +63,15 @@ const ManageItems = () => {
                         <th>Image</th>
                         <th>Bike Name</th>
                         <th>Supplier Name</th>
-                        <th>Remove Item</th>
+                        <th>Manage Item</th>
+                        <th>Update Item</th>
                     </tr>
                 </thead>
                 <tbody>
                     {
                         bikes.map(bike => <tr key={bike._id}>
                             <td>
-                                <img src={bike.img} className='custom-img' style={{ width: '20px' }} />
+                                <img src={bike.img} className='custom-img' style={{ width: '20px' }} alt='not found' />
                             </td>
                             <td>
                                 {bike.name}
@@ -53,15 +79,19 @@ const ManageItems = () => {
                             <td>
                                 {bike.supplierName}
                             </td>
-                            <td className='d-flex justify-content-around'>
-                                <button onClick={() => { removeItem(bike._id) }} className='btn btn-danger me-3'>Delete</button>
-                                <Button as={Link} to={`/inventory/${bike._id}`} variant="primary">Update</Button>
+                            <td className='d-flex justify-content-around manage-btn-container'>
+                                <button onClick={() => { removeItem(bike._id) }} className='btn manage-btn btn-danger'><TrashIcon className="h-5 w-5 text-blue-500" style={{ width: '20px' }} /></button>
+                                <button onClick={() => { addToOrder(bike) }} className='btn manage-btn btn-danger'>
+                                    <PlusIcon className="h-5 w-5 text-blue-500" style={{ width: '20px' }} />
+                                </button>
+                            </td>
+                            <td>
+                                <Button as={Link} to={`/inventory/${bike._id}`} className='manage-btn' variant="primary">Update</Button>
                             </td>
                         </tr>)
                     }
                     <tr>
                     </tr>
-
                 </tbody>
             </Table>
         </div>
